Reordered the code for QH and improved plotting.
The code for storing and plotting the evolution of the Q values is
now more flexible.
Defined Q_numQ to be the number of Q values from which we want to
see the evolution (selected randomly).
Defined Q_delta in order to store every Q_delta episode (or game).

Tested for 2x2 to print the Q values for various configurations of
QH_delta and QH_numQ.