Fixed cd2xy and xy2cd in utils used by gtp_engine.
We now give the argument n to the functinos in order to give
appropiate return for different board sizes.

Also fixed the y coordinate, since in the internal program it
is inverted (it goes from top to bottom, while in gtp syntax it
goes from bottom to top).

TODO: board implementation in a new branch.
TODO: stop using xy coordinates and instead use z position only.