Added gomill. Implemented basic play ant __str__.
Added gomill to the project to be able to see uses directly.
In board we have __str__ to print board. This was done to make
testing of play easier, because with __str__ we can print the
board and see if the play was correctly executed.

TODO: board.play() should implement suicide, ko and captures.