CNN 1st implementation complete (test more)
I created the file convolution_sutskever.py based on the MNIST
example of the tensorflow tutorials. The architecture is very
simple, but it seems to learn decently. The error on the 3 datasets
is ~70-80% meaning that it classifies correctly ~20-30% of the data,
like in the paper by Sutskever.

TODO: store parameters in files
TODO: implement player using CNN predictor