Implemented modular CNN.
Using the tutorial of CIFAR10 of tensorflow, we implemented a
CNN similar to the one implemented before.

The 9x9 datasets seems to work ok, but the 19x19 dataset does not,
so we need to test more.

TODO: Make CNN_player using the learned parameters
TODO: Test accuracy on top-k predictions