CNN Player complete and GTP engine working.
I added some functions to board to help with the board to bin
representation, also integrated the CNNPlayer to gtp_engine.py
in order to be able to play on 9x9 and 19x19 (not tested yet on
this one).

The results are promising, although we are still losing against
gnugo in 9x9. This should be because of the architecture.

TODO: modifiy dataset generator to generate 8 channels
TODO: create a many layers architecture
TODO: generate RL trainer using Policy gradient (REINFORCE)