Modifications to CNN_MODELS and dataset_generator.
The CNNs are not working very well against GnuGo. The dataset was
modified to have the same number of examples in each training file,
because there was a problem when training that made the loss=4.89 for
9x9 and loss=5.89 for 19x19.

However, it seems that the error is due to lack of weight regularization,
it seems that the weights grow too much.

TODO: try regularization